requirements.txt
scikit-learn==0.24.1
pandas==1.2.4
numpy==1.20.2
matplotlib==3.2.2
seaborn==0.10.1
tqdm==4.60.0
# http://surpriselib.com/ -> ref it
scikit-surprise==1.1.1